Pass ranges of matches to custom Matcher implementations
When the custom `Matcher` API is used, the range of each match found when
scanning a string is now passed to each matcher’s handler. That enables
more customizability, for example to be able to later replace all found
substrings within the original string.

The `scan` implementation was also a bit streamlined to accommodate this
change, since the identifier matched against now needs to be kept track of.