Enable matchers to be single-use only
This change makes it possible to create matchers that
only match a single identifier/terminator combo, and
get discarded after that.

This, combined with a new check to make sure that for
each iteration, there’s at least one eligible matcher
still remaining - can lead to some big performance
improvements if only a single piece of data is being
scanned for.