Add new APIs for scanning for a single substring
This change introduces new APIs for scanning for just a single substring
and immediately returning any result found. Using this API will be a lot
faster for use cases that only require a single match, since it won’t
require iterating through the rest of the string once a match was found.